backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

sqlalchemy>=2.0.0
sqlmodel>=0.0.16

asyncpg>=0.29.0
python-dotenv>=1.0.1
